#2 Alterando local das funções de gerenciamento de informações do perfil do usuário, e adicionando função de alteração de email.
